debug

+++ lib/libzone.h
@@ -21,12 +21,12 @@ extern ssize_t zbd_pwrite(struct zbc_device *dev, const void *buf,
 			  size_t count, uint64_t offset);
 
 extern ssize_t zbd_read_zblk(struct zbc_device *dev, void *buf, 
-                     uint32_t zoneId, uint64_t inzone_blkoff, size_t blkcnt);
+                     uint64_t zoneId, uint64_t inzone_blkoff, size_t blkcnt);
               
 extern ssize_t zbd_write_zone(struct zbc_device *dev, const void *wbuf, int force, 
-            uint32_t zoneId, uint64_t inzone_blkoff, size_t blkcnt);
+            uint64_t zoneId, uint64_t inzone_blkoff, size_t blkcnt);
 
-extern int zbd_set_wp(struct zbc_device *dev, uint32_t zoneId, uint64_t inzone_blkoff);
+extern int zbd_set_wp(struct zbc_device *dev, uint64_t zoneId, uint64_t inzone_blkoff);
 
 
 #ifdef _TEST_
